# bitlpm: Avoid allocs in CIDR trie lookups