--- a/pkg/container/bitlpm/cidr.go
+++ b/pkg/container/bitlpm/cidr.go
@@ -6,6 +6,7 @@ package bitlpm
 import (
 	"math/bits"
 	"net/netip"
+	"unsafe"
 )
 
 // CIDRTrie can hold both IPv4 and IPv6 prefixes
@@ -102,25 +103,33 @@ func (k cidrKey) Value() netip.Prefix {
 }
 
 func (k cidrKey) BitValueAt(idx uint) uint8 {
-	bytes := netip.Prefix(k).Addr().AsSlice()
-	byt := bytes[idx/8]
-	if byt&(1<<(7-(idx%8))) == 0 {
-		return 0
+	addr := netip.Prefix(k).Addr()
+	if addr.Is4() {
+		word := (*(*[2]uint64)(unsafe.Pointer(&addr)))[1]
+		return uint8((word >> (31 - idx)) & 1)
 	}
-	return 1
+	if idx < 64 {
+		word := (*(*[2]uint64)(unsafe.Pointer(&addr)))[0]
+		return uint8((word >> (63 - idx)) & 1)
+	} else {
+		word := (*(*[2]uint64)(unsafe.Pointer(&addr)))[1]
+		return uint8((word >> (127 - idx)) & 1)
+	}
 }
 
 func (k cidrKey) CommonPrefix(k2 netip.Prefix) uint {
-	out := uint(0)
-	b1 := k.Value().Addr().AsSlice()
-	b2 := k2.Addr().AsSlice()
-
-	for i := range b1 {
-		v := bits.LeadingZeros8(b1[i] ^ b2[i])
-		out += uint(v)
-		if v != 8 {
-			break
-		}
+	addr1 := netip.Prefix(k).Addr()
+	addr2 := k2.Addr()
+	words1 := (*[2]uint64)(unsafe.Pointer(&addr1))
+	words2 := (*[2]uint64)(unsafe.Pointer(&addr2))
+	if addr1.Is4() {
+		word1 := uint32((*words1)[1])
+		word2 := uint32((*words2)[1])
+		return uint(bits.LeadingZeros32(word1 ^ word2))
 	}
-	return out
+	v := bits.LeadingZeros64((*words1)[0] ^ (*words2)[0])
+	if v == 64 {
+		v += bits.LeadingZeros64((*words1)[1] ^ (*words2)[1])
+	}
+	return uint(v)
 }