CIDR key functions (pkg/container/bitlpm/cidr.go) used in lookups rely on netip.Addr.AsSlice(), which allocs a new slice. This results in many allocs during tree traversal used for all lookups in CIDR tries.

Unfortunately netip package does not provide alloc-free access to the bits needed for CIDR trie implementation. It would be preferable to extend the exported interface of bitlpm.Addr to allow for this, but in the meanwhile we can rely on the unsafe package and our own unit tests that break if the underlying implementation of the prefix bits layout changes.

This change accesses the prefix bits as an array of uint64 as stored in the underlying netip.Addr implementation. This has the main benefit of removing the need to allocate new slices whenever accessing bits in the prefix, but also making the CommonPrefix more efficient as it now operates on bigger units (uint64 vs byte).

If using unsafe with unit testing guarding against breakage is not acceptable, we could "fork" netip.Addr/Prefix implementation into a Cilium specific data types, but this would be cumbersome as well.

With this change the BenchmarkFindAffectedChildPrefixes in pkg/ipcache becomes upto 13% faster:

Before:

> pkg/ipcache$ go test -bench BenchmarkFindAffectedChildPrefixes -run XYZ -timeout 0 goos: linux
> goarch: arm64
> pkg: github.com/cilium/cilium/pkg/ipcache
> BenchmarkFindAffectedChildPrefixes/1000000/10000/Sparseness_0.330000/Trie_false-12         	      91	  12609006 ns/op	    4245 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000000/10000/Sparseness_0.330000/Trie_true-12          	  191980	      7141 ns/op	    4541 B/op	      56 allocs/op
> BenchmarkFindAffectedChildPrefixes/100000/1000/Sparseness_0.330000/Trie_false-12           	    1018	   1165072 ns/op	    4280 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/100000/1000/Sparseness_0.330000/Trie_true-12            	  284762	      3604 ns/op	    4508 B/op	      47 allocs/op
> BenchmarkFindAffectedChildPrefixes/10000/100/Sparseness_0.330000/Trie_false-12             	    9751	    122269 ns/op	    4308 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/10000/100/Sparseness_0.330000/Trie_true-12              	  512391	      2047 ns/op	    4507 B/op	      36 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000/10/Sparseness_0.330000/Trie_false-12               	   85987	     13713 ns/op	    4214 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000/10/Sparseness_0.330000/Trie_true-12                	  802904	      1411 ns/op	    4376 B/op	      26 allocs/op
> PASS
> ok  	github.com/cilium/cilium/pkg/ipcache	18.700s
> 

After:

> pkg/ipcache$ go test -bench BenchmarkFindAffectedChildPrefixes -run XYZ -timeout 0 goos: linux
> goarch: arm64
> pkg: github.com/cilium/cilium/pkg/ipcache
> BenchmarkFindAffectedChildPrefixes/1000000/10000/Sparseness_0.330000/Trie_false-12         	      94	  12744952 ns/op	    4251 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000000/10000/Sparseness_0.330000/Trie_true-12          	  194989	      6398 ns/op	    4357 B/op	      10 allocs/op
> BenchmarkFindAffectedChildPrefixes/100000/1000/Sparseness_0.330000/Trie_false-12           	    1018	   1174121 ns/op	    4280 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/100000/1000/Sparseness_0.330000/Trie_true-12            	  317127	      3222 ns/op	    4359 B/op	      10 allocs/op
> BenchmarkFindAffectedChildPrefixes/10000/100/Sparseness_0.330000/Trie_false-12             	    9727	    121524 ns/op	    4308 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/10000/100/Sparseness_0.330000/Trie_true-12              	  597798	      1772 ns/op	    4404 B/op	      10 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000/10/Sparseness_0.330000/Trie_false-12               	   84943	     13826 ns/op	    4214 B/op	       6 allocs/op
> BenchmarkFindAffectedChildPrefixes/1000/10/Sparseness_0.330000/Trie_true-12                	  858214	      1249 ns/op	    4312 B/op	      10 allocs/op
> PASS
> ok  	github.com/cilium/cilium/pkg/ipcache	15.177s
> 

Performance difference can be bigger with larger CIDR tries, as in when processing deny policies (>30% faster policy computation in a benchmark test).